Relocate project-index syntax error formatter into plugin tree
Move formatProjectIndexSyntaxError from src/shared/reporting/syntax-error-formatter.js into src/plugin/src/project-index/syntax-error-formatter.js so the project-index specific logic lives alongside the rest of the project-index helpers. Update imports, barrels, and tests to follow the new location.
