+++ src/plugin/tests/project-index-parser-errors.test.js
@@ -2,7 +2,7 @@ import assert from "node:assert/strict";
 import test from "node:test";
 
 import { getDefaultProjectIndexParser } from "../src/project-index/gml-parser-facade.js";
-import { formatProjectIndexSyntaxError } from "../../shared/reporting.js";
+import { formatProjectIndexSyntaxError } from "../src/project-index/syntax-error-formatter.js";
 
 test("project index parser reports syntax errors with context", () => {
     const parser = getDefaultProjectIndexParser();
