+++ src/shared/reporting/index.js
@@ -1,2 +1,1 @@
 export * from "./metrics-tracker.js";
-export * from "./syntax-error-formatter.js";
